Pauta 1

+++ js/cart.js
@@ -1,11 +1,16 @@
-let carritoContainer = document.getElementById('carritoContainer');
+const tipoEnvio = document.getElementById("tipoEnvio");
+const elementoCosto = document.getElementById("subtotalCosto")
+const carritoContainer = document.getElementById('carritoContainer');
+const campoCostoEnvio = document.getElementById("envioCosto");
+const campoCostoTotal = document.getElementById("totalCosto");
 let carritoActual = JSON.parse(localStorage.getItem('carrito')) || [];
+
 
 document.addEventListener("DOMContentLoaded", () => {
   mostrarInformacionEnHTML();
 });
 
-function mostrarInformacionEnHTML(data) {
+function mostrarInformacionEnHTML() {
   if (!carritoActual || carritoActual.length === 0) {
     carritoContainer.innerHTML = '<p class="alert alert-warning">El carrito está vacío</p>';
     return;
@@ -31,6 +36,7 @@ function mostrarInformacionEnHTML(data) {
       <!-- Los datos del carrito se agregarán aquí dinámicamente -->
     </tbody>
   `;
+
 
   // Obtener el cuerpo de la tabla para agregar filas
   const tbody = tabla.querySelector('tbody');
@@ -68,15 +74,20 @@ function mostrarInformacionEnHTML(data) {
     const subtotalCell = document.createElement("td");
 
     // Botón de eliminación
-    const eliminarButton = document.createElement("button");
-    eliminarButton.textContent = "Eliminar";
-    eliminarButton.classList.add("btn", "btn-danger", "btn-sm");
+    const eliminarButton = document.createElement("i");
+    eliminarButton.classList.add("btn", "custom-delete-btn", "fas", "fa-trash-alt");
     eliminarButton.dataset.productoId = producto.id;
     eliminarButton.addEventListener("click", () => {
-      const productoId = producto.id;
-      carritoActual = carritoActual.filter(item => item.id !== productoId);
-      localStorage.setItem('carrito', JSON.stringify(carritoActual));
-      mostrarInformacionEnHTML();
+
+      let productoId = producto.id;
+      var confirmacion = confirm('¿Estás seguro de que deseas eliminar este producto?');
+      // Si el usuario hace clic en "Aceptar" en la alerta de confirmación
+      if (confirmacion) {
+        carritoActual = carritoActual.filter(item => item.id !== productoId);
+        localStorage.setItem('carrito', JSON.stringify(carritoActual));
+        mostrarInformacionEnHTML();
+        mostrarCosto();
+      }
     });
 
     // Celda de la acción con el botón de eliminación
@@ -88,13 +99,10 @@ function mostrarInformacionEnHTML(data) {
       const cantidad = parseInt(cantidadInput.value);
       let subtotalValue = 0;
 
-      if (producto.currency === 'UYU') {
-        subtotalValue = cantidad * producto.unitCost * 40;
-        subtotalCell.textContent = `UYU ${subtotalValue}`;
-      } else {
-        subtotalValue = cantidad * producto.unitCost;
-        subtotalCell.textContent = `${producto.currency} ${(subtotalValue).toFixed(2)}`;
-      }
+
+      subtotalValue = cantidad * producto.unitCost;
+      subtotalCell.textContent = ` ${(subtotalValue).toFixed(2)}`;
+
 
       // Actualizar el producto en el carritoActual con la nueva cantidad
       const productoIndex = carritoActual.findIndex(item => item.id === producto.id);
@@ -132,13 +140,52 @@ function mostrarInformacionEnHTML(data) {
 
 function mostrarCosto() {
   let subtotal = 0;
+
   carritoActual.forEach(producto => {
     if (producto.currency === 'UYU') {
-      subtotal += producto.count * producto.unitCost * 40;
+      subtotal += producto.count * producto.unitCost / 40;
     } else {
       subtotal += producto.count * producto.unitCost;
     }
+
   });
+
+
+
+  elementoCosto.textContent = `${parseFloat(subtotal).toFixed(2)}`;
+
+
+  function obtenerTipodeEnvio() {
+    let valorSeleccionado = tipoEnvio.value;
+    // Define las tarifas de envío para cada tipo
+    const tarifasEnvio = {
+      premium: 0.15,   // 15%
+      express: 0.07,   // 7%
+      estandar: 0.05   // 5%
+    };
+    return tarifasEnvio[valorSeleccionado];
+  }
+
+
+  function calcularCosto() {
+    let valorSeleccionado = obtenerTipodeEnvio() ?? 0
+    let costoEnvio = subtotal * valorSeleccionado;
+    campoCostoEnvio.textContent = `${parseFloat(costoEnvio).toFixed(2)}`;
+    calcularTotal();
+
+  }
+
+
+  function calcularTotal() {
+    let total = parseFloat(elementoCosto.textContent) + parseFloat(campoCostoEnvio.textContent);
+    campoCostoTotal.textContent = total.toFixed(2);
+  }
+  calcularTotal();
+
+  campoCostoEnvio.addEventListener("change", calcularCosto());
+  tipoEnvio.addEventListener("change", () => calcularCosto());
+
+
   document.getElementById("subtotalCosto").textContent = `${parseFloat(subtotal).toFixed(2)}`;
 }
 
@@ -238,4 +285,5 @@ function disableInputs(section) {
   for (const input of inputs) {
       input.setAttribute("disabled", "true");
   }
+
 }
